refactor: 리뷰 상세 조회 분리된 api 답변 페이지에 적용
