+++ src/pages/ReviewReplyPage/components/ReviewReplyStart/ReviewSelect.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useState, Dispatch, SetStateAction } from 'react'
 import { SubmitHandler, useFieldArray, useFormContext } from 'react-hook-form'
 import { Profile, SearchBar } from '@/components'
 import { useUser } from '@/apis/hooks'
-import { Question } from '@/apis/hooks/useGetReviewFirst'
+import { Question } from '@/apis/hooks/useGetReviewForCreator'
 import { CloseIcon } from '@/assets/icons'
 import { ReviewReplyStartType } from '../../types'
