+++ src/pages/ReviewReplyPage/hooks/useClickNextButton.ts
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from 'react'
-import { Question, Receiver } from '@/apis/hooks/useGetReviewFirst'
+import { Question, Receiver } from '@/apis/hooks/useGetReviewForCreator'
 
 interface UseClickNextButtonProps {
   questions: Question[]
